Create element_tyes and value_containers.
