+++ src/compute/compute_kernels.cpp
@@ -9,4 +9,3 @@ template class specfem::compute::impl::kernels::material_kernels<
 template class specfem::compute::impl::kernels::material_kernels<
     specfem::element::medium_tag::acoustic,
     specfem::element::property_tag::isotropic>;
-
