+++ src/compute/compute_mesh.cpp
@@ -228,14 +228,14 @@ specfem::compute::mesh_to_compute_mapping::mesh_to_compute_mapping(
 
   const int nspec = tags.nspec;
 
-  constexpr auto element_types = specfem::element::element_types();
-  constexpr int total_element_types = element_types.size();
+  constexpr auto medium_tags = specfem::element::medium_tags();
+  constexpr int total_medium_tags = medium_tags.size();
 
-  std::array<std::vector<int>, total_element_types> element_type_ispec;
+  std::array<std::vector<int>, total_medium_tags> element_type_ispec;
   int total_counted = 0;
 
-  for (int i = 0; i < total_element_types; i++) {
-    const auto [medium_tag, property_tag, boundary_tag] = element_types[i];
+  for (int i = 0; i < total_medium_tags; i++) {
+    const auto [medium_tag, property_tag, boundary_tag] = medium_tags[i];
     for (int ispec = 0; ispec < nspec; ispec++) {
       const auto tag = tags.tags_container(ispec);
       if (tag.medium_tag == medium_tag && tag.property_tag == property_tag &&
