+++ src/compute/compute_properties.cpp
@@ -11,9 +11,9 @@ void compute_number_of_elements_per_medium(
     const int nspec, const specfem::compute::mesh_to_compute_mapping &mapping,
     const specfem::mesh::tags<specfem::dimension::type::dim2> &tags,
     const specfem::kokkos::HostView1d<specfem::element::medium_tag>
-        &h_element_types,
+        &h_medium_tags,
     const specfem::kokkos::HostView1d<specfem::element::property_tag>
-        &h_element_property,
+        &h_property_tags,
     int &n_elastic_isotropic, int &n_elastic_anisotropic, int &n_acoustic) {
 
   Kokkos::parallel_reduce(
@@ -24,16 +24,16 @@ void compute_number_of_elements_per_medium(
         const int ispec_mesh = mapping.compute_to_mesh(ispec);
         if (tags.tags_container(ispec_mesh).medium_tag ==
             specfem::element::medium_tag::elastic) {
-          h_element_types(ispec) = specfem::element::medium_tag::elastic;
+          h_medium_tags(ispec) = specfem::element::medium_tag::elastic;
           if (tags.tags_container(ispec_mesh).property_tag ==
               specfem::element::property_tag::isotropic) {
             n_elastic_isotropic++;
-            h_element_property(ispec) =
+            h_property_tags(ispec) =
                 specfem::element::property_tag::isotropic;
           } else if (tags.tags_container(ispec_mesh).property_tag ==
                      specfem::element::property_tag::anisotropic) {
             n_elastic_anisotropic++;
-            h_element_property(ispec) =
+            h_property_tags(ispec) =
                 specfem::element::property_tag::anisotropic;
           } else {
             throw std::runtime_error("Unknown property tag");
@@ -41,10 +41,10 @@ void compute_number_of_elements_per_medium(
         } else if (tags.tags_container(ispec_mesh).medium_tag ==
                    specfem::element::medium_tag::acoustic) {
           n_acoustic++;
-          h_element_types(ispec) = specfem::element::medium_tag::acoustic;
+          h_medium_tags(ispec) = specfem::element::medium_tag::acoustic;
           if (tags.tags_container(ispec_mesh).property_tag ==
               specfem::element::property_tag::isotropic) {
-            h_element_property(ispec) =
+            h_property_tags(ispec) =
                 specfem::element::property_tag::isotropic;
           } else {
             throw std::runtime_error("Unknown property tag");
@@ -67,12 +67,12 @@ specfem::compute::properties::properties(
     const specfem::mesh::tags<specfem::dimension::type::dim2> &tags,
     const specfem::mesh::materials &materials)
     : nspec(nspec), ngllz(ngllz), ngllx(ngllx),
-      element_types("specfem::compute::properties::element_types", nspec),
-      h_element_types(Kokkos::create_mirror_view(element_types)),
+      medium_tags("specfem::compute::properties::medium_tags", nspec),
+      h_medium_tags(Kokkos::create_mirror_view(medium_tags)),
       property_index_mapping(
           "specfem::compute::properties::property_index_mapping", nspec),
-      element_property("specfem::compute::properties::element_property", nspec),
-      h_element_property(Kokkos::create_mirror_view(element_property)),
+      property_tags("specfem::compute::properties::property_tags", nspec),
+      h_property_tags(Kokkos::create_mirror_view(property_tags)),
       h_property_index_mapping(
           Kokkos::create_mirror_view(property_index_mapping)) {
 
@@ -81,8 +81,8 @@ specfem::compute::properties::properties(
   int n_elastic_anisotropic;
   int n_acoustic;
 
-  compute_number_of_elements_per_medium(nspec, mapping, tags, h_element_types,
-                                        h_element_property, n_elastic_isotropic,
+  compute_number_of_elements_per_medium(nspec, mapping, tags, h_medium_tags,
+                                        h_property_tags, n_elastic_isotropic,
                                         n_elastic_anisotropic, n_acoustic);
 
   acoustic_isotropic = specfem::compute::impl::properties::material_property<
@@ -104,8 +104,8 @@ specfem::compute::properties::properties(
       h_property_index_mapping);
 
   Kokkos::deep_copy(property_index_mapping, h_property_index_mapping);
-  Kokkos::deep_copy(element_types, h_element_types);
-  Kokkos::deep_copy(element_property, h_element_property);
+  Kokkos::deep_copy(medium_tags, h_medium_tags);
+  Kokkos::deep_copy(property_tags, h_property_tags);
 
   return;
 }
@@ -118,7 +118,7 @@ specfem::compute::properties::get_elements_on_host(
 
   int nelements = 0;
   for (int ispec = 0; ispec < nspec; ispec++) {
-    if (h_element_types(ispec) == medium) {
+    if (h_medium_tags(ispec) == medium) {
       nelements++;
     }
   }
@@ -128,7 +128,7 @@ specfem::compute::properties::get_elements_on_host(
 
   nelements = 0;
   for (int ispec = 0; ispec < nspec; ispec++) {
-    if (h_element_types(ispec) == medium) {
+    if (h_medium_tags(ispec) == medium) {
       elements(nelements) = ispec;
       nelements++;
     }
@@ -165,8 +165,8 @@ specfem::compute::properties::get_elements_on_host(
 
   int nelements = 0;
   for (int ispec = 0; ispec < nspec; ispec++) {
-    if (h_element_types(ispec) == medium &&
-        h_element_property(ispec) == property) {
+    if (h_medium_tags(ispec) == medium &&
+        h_property_tags(ispec) == property) {
       nelements++;
     }
   }
@@ -176,8 +176,8 @@ specfem::compute::properties::get_elements_on_host(
 
   nelements = 0;
   for (int ispec = 0; ispec < nspec; ispec++) {
-    if (h_element_types(ispec) == medium &&
-        h_element_property(ispec) == property) {
+    if (h_medium_tags(ispec) == medium &&
+        h_property_tags(ispec) == property) {
       elements(nelements) = ispec;
       nelements++;
     }
