+++ tests/unit-tests/assembly/compute_wavefield/compute_wavefield.cpp
@@ -20,8 +20,8 @@ void test_element_wavefield(
 
   const auto properties = assembly.properties;
 
-  const auto medium = properties.h_element_types(ispec);
-  const auto property = properties.h_element_property(ispec);
+  const auto medium = properties.h_medium_tags(ispec);
+  const auto property = properties.h_property_tags(ispec);
 
   if ((medium == specfem::element::medium_tag::elastic) &&
       (property == specfem::element::property_tag::isotropic)) {
